Move perf dialog method to main window
